+++ tests/errors/semantic/blocking/EXTEND_LIST_INCOMPATIBLE_TYPES_2.py
@@ -1,0 +1,7 @@
+# pylint: disable=missing-function-docstring, missing-module-docstring
+
+a = [1,2,3]
+b = (((4,5), (3, 4)))
+a.extend(b)
+print(a)
+